fastAPI-example/pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = ""
authors = [
    {name = "Henrik",email = "602119@stud.hvl.no"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "dynamic-frcm (>=1.0.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
